+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
-
-* TODO: Describe initial release.
+Initial commit
+## 0.0.2
+Fixed unhandled exception on login
+## 0.0.3
+Added Bagrut grades and Hatamot APIs.
+## 0.0.4
+Added Api enum and Data Processor to Api Controller.